Project 1 - Not-Hot-Dog/model/requirements.txt
numpy==1.18.1
pandas==1.0.5
tensorflow-cpu==2.1.0
h5py==2.10.0